requirements.txt
africastalking==1.2.3
appdirs==1.4.4
astroid==2.4.2
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
confusable-homoglyphs==3.2.0
contextlib2==0.6.0.post1
crypto==1.4.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2
django-allauth==0.39.1
django-braces==1.14.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-generic-positions==0.2.3
django-heroku==0.3.1
django-hvad==1.8.0
django-libs==2.0.3
django-model-utils==4.0.0
django-registration==3.1
django-review==1.10.0
django-star-ratings==0.9.1
django-user-media==1.2.5
easy-thumbnails==2.7
filelock==3.0.12
Flutterwave==1.0.7
future==0.18.2
graphviz==0.14.1
gunicorn==20.0.4
idna==2.8
isort==5.5.2
joblib==0.16.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Naked==0.1.31
numpy==1.19.2
oauthlib==3.0.1
onnx==1.7.0
pep8==1.7.1
Pillow==6.2.2
protobuf==3.13.0
psycopg2==2.8.6
pycodestyle==2.5.0
pycrypto==2.6.1
pydot==1.4.1
pylint==2.6.0
pyparsing==2.4.7
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
PyYAML==5.3.1
requests==2.21.0
requests-oauthlib==1.2.0
schema==0.7.2
scikit-learn==0.23.2
scipy==1.5.2
shellescape==3.8.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.2.4
stripe==2.27.0
swapper==1.1.2.post1
threadpoolctl==2.1.0
toml==0.10.1
torch==1.6.0+cpu
torchvision==0.7.0+cpu
torchviz==0.0.1
tqdm==4.49.0
typing-extensions==3.7.4.3
urllib3==1.24.2
virtualenv==20.0.31
whitenoise==5.2.0
wrapt==1.12.1
